fixing some gcc warnings
git-svn-id: https://local.ugene.unipro.ru/svn/ugene/trunk@849 7ca3a678-163c-e011-b84f-001e8c0267ff

+++ src/plugins/genome_aligner/src/GenomeAlignerIndexPart.cpp
@@ -96,7 +96,7 @@ bool IndexPart::load(int part) {
     char *buff = (char*)sArray;
     size = partFiles[part]->read(buff, saLengths[currentPart]*sizeof(SAType));
     assert(size == saLengths[currentPart]*sizeof(SAType));
-    if (size != saLengths[currentPart]*sizeof(SAType)) {
+    if (size != qint64(saLengths[currentPart]) * sizeof(SAType)) {
         return false;
     }
